Better locking logic in block cache
